refactor(transport/rclcpp): move manual_control_node.hpp into rclcpp transport for structural symmetry
